move utils and const to separate files
